requirements.txt
Flask>=3.0.0
Pillow>=10.0.0
PyYAML>=6.0.0
ultralytics>=8.0.0
